split convert module between request/response
